Backend for US14, 15 and 17

+++ app/Http/Controllers/OrderControllerAPI.php
@@ -6,25 +6,41 @@ use App\Http\Resources\Order as OrderResource;
 
 use Illuminate\Http\Request;
 use App\Order;
+use Illuminate\Support\Facades\Auth;
 
 class OrderControllerAPI extends Controller
 {
     public function index(Request $request)
     {
-        return OrderResource::collection(Order::paginate(20));
+        $uid = Auth::user()->id;
+        return OrderResource::collection(Order::join('meals', 'orders.meal_id', '=', 'meals.id')->where('meals.responsible_waiter_id', $uid)->orderBy('orders.updated_at', 'desc')->paginate(50));
     }
 
-    public function waiter(Request $request)
-    {
-        $uid = Auth::user()->id;
-        return OrderResource::collection(Order::where('meal.responsible_waiter_id', $uid)->where('state', 'pending')->orWhere('state', 'confirmed')->paginate(20));
-    }
+// Trash code
+//    public function waiter(Request $request)
+//    {
+//        $uid = Auth::user()->id;
+//        return OrderResource::collection(Order::join('meals', 'orders.meal_id', '=', 'meals.id')->where('meals.responsible_waiter_id', $uid)->where('meals.state', 'active')->where('orders.state', 'pending')->orWhere('orders.state', 'confirmed')->paginate(20));
+//    }
 
     public function show($id)
     {
         return new OrderResource(Order::find($id));
     }
 
+    public function deliver($id)
+    {
+        $uid = Auth::user()->id;
+        $order = Order::join('meals', 'orders.meal_id', '=', 'meals.id')->where('meals.responsible_waiter_id', $uid)->where('orders.state', 'prepared')->findOrFail($id);
+
+        $order->state = 'delivered';
+        $order->save();
+
+        return response()->json(null, 200);
+    }
+
+// TODO: Diogo's Task
+//
 //    public function store(Request $request)
 //    {
 //        $request->validate([
@@ -39,23 +55,17 @@ class OrderControllerAPI extends Controller
 //        $user->save();
 //        return response()->json(new UserResource($user), 201);
 //    }
-//
-//    public function update(Request $request, $id)
-//    {
-//        $request->validate([
-//            'name' => 'required|min:3|regex:/^[A-Za-záàâãéèêíóôõúçÁÀÂÃÉÈÍÓÔÕÚÇ ]+$/',
-//            'email' => 'required|email|unique:users,email,'.$id,
-//            'age' => 'integer|between:18,75'
-//        ]);
-//        $user = User::findOrFail($id);
-//        $user->update($request->all());
-//        return new UserResource($user);
-//    }
 
     public function destroy($id)
     {
-        $order = Order::findOrFail($id);
-        $order->delete();
-        return response()->json(null, 204);
+        $order = Order::where('state', 'pending')->findOrFail($id);
+        $dtime = Carbon\Carbon::now()->subSeconds(5);
+
+        if($dtime < $order->created_at){
+            $order->delete();
+            return response()->json(null, 204);
+        }
+
+        return response()->json(null, 406);
     }
 }
